refactor: comment apis features => entities로 이동
